+++ cypress/support/commands.js
@@ -23,3 +23,21 @@
 //
 // -- This is will overwrite an existing command --
 // Cypress.Commands.overwrite("visit", (originalFn, url, options) => { ... })
+
+Cypress.Commands.add('filloutFormElement', (selector, content, position) => {
+
+  switch (position) {
+    case undefined:
+      cy.get(selector).type(content);
+      break;
+    case 'first':
+      cy.get(selector).first().type(content);
+      break;
+    case 'last':
+      cy.get(selector).last().type(content);
+      break;
+    default:
+      cy.get(selector).eq(position).type(content);
+      break;
+  }
+});